Split tests/integration/test_integration.py into per-axis test modules
The file had grown to 2718 lines covering eleven distinct axes (basic
golden output, combined variable forms, format variants, line-ending
variants, heterogeneous strategies, pre-indent level, calls, call
variants, orphan goldens, format-enumeration properties, and a meta-test
of the in-memory golden assertion). Each axis now lives in its own test
file, with shared scaffolding split into focused helper modules
(language_specs, golden_assertions, variant_cases, case_discovery,
call_cases, call_variant_cases, variable_form_wrapping). The cases_dir
fixture moves into conftest.py.
