+++ tests/integration/conftest.py
@@ -1,0 +1,13 @@
+"""Fixtures shared across the integration test modules."""
+
+from pathlib import Path
+
+import pytest
+
+
+@pytest.fixture(name="cases_dir")
+def fixture_cases_dir(request: pytest.FixtureRequest) -> Path:
+    """Return the absolute path to the integration test cases
+    directory.
+    """
+    return request.config.rootpath / "tests" / "integration" / "cases"
