+++ tests/integration/test_format_enumeration_properties.py
@@ -1,0 +1,54 @@
+"""Property checks that every language exposes its formatter
+enumerations.
+"""
+
+import enum
+
+import pytest
+
+import literalizer
+
+from .language_specs import sorted_languages
+
+
+@pytest.mark.parametrize(
+    argnames="language_cls",
+    argvalues=sorted_languages(),
+    ids=[c.__name__ for c in sorted_languages()],
+)
+def test_format_enumeration_properties(
+    language_cls: literalizer.LanguageCls,
+) -> None:
+    """Every language exposes iterable format-enumeration properties."""
+    spec = language_cls()
+    assert issubclass(spec.bytes_formats, enum.Enum)
+    assert len(spec.bytes_formats) >= 1
+    assert issubclass(spec.sequence_formats, enum.Enum)
+    assert len(spec.sequence_formats) >= 1
+    assert issubclass(spec.set_formats, enum.Enum)
+    assert len(spec.set_formats) >= 1
+    assert issubclass(spec.date_formats, enum.Enum)
+    assert len(spec.date_formats) >= 1
+    assert issubclass(spec.datetime_formats, enum.Enum)
+    assert len(spec.datetime_formats) >= 1
+    assert issubclass(spec.comment_formats, enum.Enum)
+    assert len(spec.comment_formats) >= 1
+    assert issubclass(spec.declaration_styles, enum.Enum)
+    assert len(spec.declaration_styles) >= 1
+    assert issubclass(spec.dict_formats, enum.Enum)
+    assert len(spec.dict_formats) >= 1
+    assert issubclass(spec.float_formats, enum.Enum)
+    assert len(spec.float_formats) >= 1
+    assert issubclass(spec.integer_formats, enum.Enum)
+    assert len(spec.integer_formats) >= 1
+    assert issubclass(spec.numeric_separators, enum.Enum)
+    assert len(spec.numeric_separators) >= 1
+    assert issubclass(spec.numeric_styles, enum.Enum)
+    assert len(spec.numeric_styles) >= 1
+    assert issubclass(spec.string_formats, enum.Enum)
+    assert len(spec.string_formats) >= 1
+    assert issubclass(spec.trailing_commas, enum.Enum)
+    assert len(spec.trailing_commas) >= 1
+    assert issubclass(spec.line_endings, enum.Enum)
+    assert len(spec.line_endings) >= 1
+    assert issubclass(spec.call_styles, enum.Enum)
